Created new module repo-http-models. Moved ModelId, ModelResource and ModelType from repo-core to repo-http-models. Fixed the dependecies in repo-core and repo-ui due to this change
