+++ server/repo/repo-core/src/main/java/org/eclipse/vorto/repository/internal/service/IModelParser.java
@@ -16,7 +16,7 @@ package org.eclipse.vorto.repository.internal.service;
 
 import java.io.InputStream;
 
-import org.eclipse.vorto.repository.model.ModelResource;
+import org.eclipse.vorto.http.model.ModelResource;
 
 /**
  * @author Alexander Edelmann - Robert Bosch (SEA) Pte. Ltd.
