+++ server/repo/repo-core/src/main/java/org/eclipse/vorto/repository/internal/service/emf/AbstractModelParser.java
@@ -23,11 +23,11 @@ import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.common.util.URI;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.vorto.core.api.model.model.Model;
+import org.eclipse.vorto.http.model.ModelId;
+import org.eclipse.vorto.http.model.ModelResource;
+import org.eclipse.vorto.http.model.ModelType;
 import org.eclipse.vorto.repository.internal.service.IModelParser;
 import org.eclipse.vorto.repository.model.ModelEMFResource;
-import org.eclipse.vorto.repository.model.ModelId;
-import org.eclipse.vorto.repository.model.ModelResource;
-import org.eclipse.vorto.repository.model.ModelType;
 import org.eclipse.vorto.repository.validation.ValidationException;
 import org.eclipse.xtext.linking.impl.XtextLinkingDiagnostic;
 import org.eclipse.xtext.resource.XtextResource;
