+++ server/repo/repo-core/src/main/java/org/eclipse/vorto/repository/internal/service/validation/BulkModelReferencesValidation.java
@@ -19,9 +19,9 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.eclipse.vorto.http.model.ModelId;
+import org.eclipse.vorto.http.model.ModelResource;
 import org.eclipse.vorto.repository.internal.service.validation.exception.CouldNotResolveReferenceException;
-import org.eclipse.vorto.repository.model.ModelId;
-import org.eclipse.vorto.repository.model.ModelResource;
 import org.eclipse.vorto.repository.service.IModelRepository;
 import org.eclipse.vorto.repository.validation.ValidationException;
