+++ server/repo/repo-core/src/main/java/org/eclipse/vorto/repository/model/Comment.java
@@ -24,6 +24,8 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
 
+import org.eclipse.vorto.http.model.ModelId;
+
 /**
  * @author Alexander Edelmann - Robert Bosch (SEA) Pte. Ltd.
  */
