+++ server/repo/repo-core/src/main/java/org/eclipse/vorto/repository/model/ModelEMFResource.java
@@ -26,6 +26,9 @@ import org.eclipse.vorto.core.api.model.informationmodel.InformationModel;
 import org.eclipse.vorto.core.api.model.mapping.MappingModel;
 import org.eclipse.vorto.core.api.model.model.Model;
 import org.eclipse.vorto.core.api.model.model.ModelReference;
+import org.eclipse.vorto.http.model.ModelId;
+import org.eclipse.vorto.http.model.ModelResource;
+import org.eclipse.vorto.http.model.ModelType;
 
 /**
  * @author Alexander Edelmann - Robert Bosch (SEA) Pte. Ltd.
