+++ server/repo/repo-core/src/main/java/org/eclipse/vorto/repository/notification/message/CheckinMessage.java
@@ -14,11 +14,11 @@
  *******************************************************************************/
 package org.eclipse.vorto.repository.notification.message;
 
-import org.eclipse.vorto.repository.model.ModelResource;
-import org.eclipse.vorto.repository.model.User;
-
 import java.util.HashMap;
 import java.util.Map;
+
+import org.eclipse.vorto.http.model.ModelResource;
+import org.eclipse.vorto.repository.model.User;
 
 public class CheckinMessage extends AbstractMessage {
