+++ server/repo/repo-core/src/main/java/org/eclipse/vorto/repository/service/IGeneratorService.java
@@ -17,9 +17,9 @@ package org.eclipse.vorto.repository.service;
 import java.util.Collection;
 import java.util.Set;
 
+import org.eclipse.vorto.http.model.ModelId;
 import org.eclipse.vorto.repository.model.GeneratedOutput;
 import org.eclipse.vorto.repository.model.GeneratorServiceInfo;
-import org.eclipse.vorto.repository.model.ModelId;
 import org.eclipse.vorto.repository.model.ServiceClassifier;
 
 /**
