+++ server/repo/repo-core/src/main/java/org/eclipse/vorto/repository/service/IModelRepository.java
@@ -16,8 +16,8 @@ package org.eclipse.vorto.repository.service;
 
 import java.util.List;
 
-import org.eclipse.vorto.repository.model.ModelId;
-import org.eclipse.vorto.repository.model.ModelResource;
+import org.eclipse.vorto.http.model.ModelId;
+import org.eclipse.vorto.http.model.ModelResource;
 import org.eclipse.vorto.repository.model.UploadModelResult;
 
 /**
