+++ server/repo/repo-core/src/test/java/org/eclipse/vorto/repository/service/NotificationServiceTest.java
@@ -19,9 +19,9 @@ import static org.junit.Assert.assertTrue;
 
 import java.util.Date;
 
-import org.eclipse.vorto.repository.model.ModelId;
-import org.eclipse.vorto.repository.model.ModelResource;
-import org.eclipse.vorto.repository.model.ModelType;
+import org.eclipse.vorto.http.model.ModelId;
+import org.eclipse.vorto.http.model.ModelResource;
+import org.eclipse.vorto.http.model.ModelType;
 import org.eclipse.vorto.repository.model.User;
 import org.eclipse.vorto.repository.notification.EmailNotificationService;
 import org.eclipse.vorto.repository.notification.message.CheckinMessage;
