+++ server/repo/repo-http-models/src/main/java/org/eclipse/vorto/http/model/ModelType.java
@@ -12,7 +12,8 @@
  * Contributors:
  * Bosch Software Innovations GmbH - Please refer to git log
  *******************************************************************************/
-package org.eclipse.vorto.repository.model;
+package org.eclipse.vorto.http.model;
+
 
 /**
  * @author Alexander Edelmann - Robert Bosch (SEA) Pte. Ltd.
