+++ server/repo/repo-ui/pom.xml
@@ -19,6 +19,11 @@
 			<artifactId>repo-core</artifactId>
 			<version>${project.version}</version>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.vorto</groupId>
+			<artifactId>repo-http-models</artifactId>
+			<version>${project.version}</version>
+		</dependency>
 
 		<!-- Spring Boot dependencies -->
 		<dependency>
@@ -91,7 +96,7 @@
 				</exclusion>
 			</exclusions>
 		</dependency>
-		
+
 		<dependency>
 			<groupId>org.webjars.bower</groupId>
 			<artifactId>angular-swagger-ui</artifactId>
@@ -110,7 +115,7 @@
 					<artifactId>angular</artifactId>
 				</exclusion>
 			</exclusions>
-		</dependency>		
+		</dependency>
 
 		<!-- Database -->
