+++ server/repo/repo-ui/src/main/java/org/eclipse/vorto/repository/web/comment/CommentController.java
@@ -18,8 +18,8 @@ import java.util.List;
 
 import javax.validation.Valid;
 
+import org.eclipse.vorto.http.model.ModelId;
 import org.eclipse.vorto.repository.model.Comment;
-import org.eclipse.vorto.repository.model.ModelId;
 import org.eclipse.vorto.repository.service.ICommentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -35,8 +35,8 @@ import org.springframework.web.bind.annotation.RestController;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponses;
 import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
 
 /**
  * @author Alexander Edelmann - Robert Bosch (SEA) Pte. Ltd.
